requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
pypdf2>=3.0.0
pdfplumber>=0.9.0
python-docx>=0.8.11
spacy>=3.6.0
nltk>=3.8.0
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0
openai>=1.3.0
langchain>=0.0.300
chromadb>=0.4.0
faiss-cpu>=1.7.4
sqlalchemy>=2.0.0
pymongo>=4.5.0
reportlab>=4.0.0
jinja2>=3.1.0
autogen>=0.2.0